+++ Changelog.md
@@ -3,8 +3,9 @@
 ## 0.9.2
 - Benchmarks
 - implement `Entropy` and `KlDivergence` for `Poisson`
-- Implement `QuadBounds` for `Mixture<Poisson>`
+- Implement `QuadBounds` and `Entropy` for `Mixture<Poisson>`
 - Implement `Mean` for mixtures of distributions that implement `Mean<f64>` or `Mean<f32>`
+- `Rv` trait has `sample_stream` method that produces a never-ending iterator of random numbers.
 
 ## 0.9.1
 - Remove printlns